fix(reliability): readiness endpoint no longer does cascading checks
The purpose of the readiness endpoint is to signal that the service is
ready to accept requests. This endpoint is often used by Kubernetes to
detect a container will back a service. In the Kubernetes context, it
is undesirable to have the full array of health checks part of the
readiness probe. As such, this will no longer be the default behavior.